+++ components/collector/tests/source_collectors/dependency_track/test_source_up_to_dateness.py
@@ -4,17 +4,20 @@ from datetime import datetime, timedelta
 
 from dateutil.tz import tzlocal
 
-from source_collectors.dependency_track.base import DependencyTrackBase, DependencyTrackProject
+from source_collectors.dependency_track.base import (
+    DependencyTrackBase,
+    DependencyTrackMetrics,
+    DependencyTrackProject,
+)
 
-from tests.source_collectors.source_collector_test_case import SourceCollectorTestCase
+from .base_test import DependencyTrackTestCase
 
 
-class DependencyTrackSourceUpToDatenessVersionTest(SourceCollectorTestCase):
+class DependencyTrackSourceUpToDatenessVersionTest(DependencyTrackTestCase):
     """Unit tests for the Dependency-Track source up-to-dateness collector."""
 
     METRIC_ADDITION = "min"
     METRIC_TYPE = "source_up_to_dateness"
-    SOURCE_TYPE = "dependency_track"
     LANDING_URL = "https://dependency_track"
 
     def projects(self) -> list[DependencyTrackProject]:
@@ -22,11 +25,27 @@ class DependencyTrackSourceUpToDatenessVersionTest(SourceCollectorTestCase):
         now = datetime.now(tz=tzlocal()).replace(microsecond=0)
         self.yesterday = now - timedelta(days=1)
         yesterday_timestamp = round(self.yesterday.timestamp() * 1000)
+        self.day_before_yesterday = now - timedelta(days=2)
+        day_before_yesterday_timestamp = round(self.day_before_yesterday.timestamp() * 1000)
         self.last_week = now - timedelta(days=7)
-        last_week_timestamp = round(self.last_week.timestamp() * 1000)
+        self.last_week_timestamp = round(self.last_week.timestamp() * 1000)
+        self.few_days_ago = now - timedelta(days=4)
+        self.few_days_ago_timestamp = round(self.few_days_ago.timestamp() * 1000)
         return [
-            DependencyTrackProject(name="Project 1", version="1.1", uuid="p1", lastBomImport=yesterday_timestamp),
-            DependencyTrackProject(name="Project 2", version="1.2", uuid="p2", lastBomImport=last_week_timestamp),
+            DependencyTrackProject(
+                name="Project 1",
+                version="1.1",
+                uuid="p1",
+                lastBomImport=yesterday_timestamp,
+                metrics=DependencyTrackMetrics(lastOccurrence=self.few_days_ago_timestamp),
+            ),
+            DependencyTrackProject(
+                name="Project 2",
+                version="1.2",
+                uuid="p2",
+                lastBomImport=self.last_week_timestamp,
+                metrics=DependencyTrackMetrics(lastOccurrence=day_before_yesterday_timestamp),
+            ),
         ]
 
     def entities(self) -> list[dict[str, str]]:
@@ -34,12 +53,14 @@ class DependencyTrackSourceUpToDatenessVersionTest(SourceCollectorTestCase):
         return [
             {
                 "key": "p1",
+                "last_bom_analysis": self.few_days_ago.isoformat(),
                 "last_bom_import": self.yesterday.isoformat(),
                 "project": "Project 1",
                 "project_landing_url": "/projects/p1",
             },
             {
                 "key": "p2",
+                "last_bom_analysis": self.day_before_yesterday.isoformat(),
                 "last_bom_import": self.last_week.isoformat(),
                 "project": "Project 2",
                 "project_landing_url": "/projects/p2",
@@ -50,6 +71,54 @@ class DependencyTrackSourceUpToDatenessVersionTest(SourceCollectorTestCase):
         """Test that the source up-to-dateness can be measured."""
         response = await self.collect(get_request_json_return_value=self.projects())
         self.assert_measurement(response, value="7", landing_url=self.LANDING_URL, entities=self.entities())
+
+    async def test_source_up_to_dateness_missing_bom_analysis(self):
+        """Test that the source up-to-dateness can be measured even if the BOM has no last analysis datetime."""
+        projects = self.projects()
+        projects.append(
+            DependencyTrackProject(
+                name="Project 3",
+                version="1.3",
+                uuid="p3",
+                lastBomImport=self.last_week_timestamp,
+                metrics=DependencyTrackMetrics(),
+            )
+        )
+        entities = self.entities()
+        entities.append(
+            {
+                "key": "p3",
+                "last_bom_analysis": "",
+                "last_bom_import": self.last_week.isoformat(),
+                "project": "Project 3",
+                "project_landing_url": "/projects/p3",
+            }
+        )
+        response = await self.collect(get_request_json_return_value=projects)
+        self.assert_measurement(response, value="7", landing_url=self.LANDING_URL, entities=entities)
+
+    async def test_source_up_to_dateness_missing_bom_import(self):
+        """Test that the source up-to-dateness can be measured even if the BOM has no import datetime."""
+        projects = self.projects()
+        projects.append(
+            DependencyTrackProject(
+                name="Project 3",
+                version="1.3",
+                uuid="p3",
+            )
+        )
+        entities = self.entities()
+        entities.append(
+            {
+                "key": "p3",
+                "last_bom_analysis": "",
+                "last_bom_import": "",
+                "project": "Project 3",
+                "project_landing_url": "/projects/p3",
+            }
+        )
+        response = await self.collect(get_request_json_return_value=projects)
+        self.assert_measurement(response, value="7", landing_url=self.LANDING_URL, entities=entities)
 
     async def test_filter_by_project_name(self):
         """Test that projects can be filtered by name."""
@@ -75,6 +144,18 @@ class DependencyTrackSourceUpToDatenessVersionTest(SourceCollectorTestCase):
         self.set_source_parameter("project_versions", ["1.3", "1.4"])
         response = await self.collect(get_request_json_return_value=self.projects())
         self.assert_measurement(response, parse_error="No projects found")
+
+    async def test_filter_by_event_type_last_bom_import(self):
+        """Test that projects can be filtered by event type."""
+        self.set_source_parameter("project_event_types", ["last BOM import"])
+        response = await self.collect(get_request_json_return_value=self.projects())
+        self.assert_measurement(response, value="7", landing_url=self.LANDING_URL, entities=self.entities())
+
+    async def test_filter_by_event_type_last_bom_analysis(self):
+        """Test that projects can be filtered by event type."""
+        self.set_source_parameter("project_event_types", ["last BOM analysis"])
+        response = await self.collect(get_request_json_return_value=self.projects())
+        self.assert_measurement(response, value="4", landing_url=self.LANDING_URL, entities=self.entities())
 
     async def test_source_up_to_dateness_with_pagination(self):
         """Test that the source up-to-dateness can be measured when pagination is needed."""